// src/app/models/transaction-global.ts
export class Transaction {
    id: number;
    clerc: string; 
    client: string;
    sellerName: string; // nom du vendeur
    buyerName: string; // nom de l'acquéreur
    RepoName: string; // nom du dossier/ Portefeuille
    address: string; // adresse du bien
    signatureDate: string;
    dataRoomCreator: string; // confrère ayant réalisé la data room
    typeDR: string;
  }
  